dx_proj_onto_plane_at_0

+++ crates/sophus_lie/src/groups/translation_product_product.rs
@@ -473,10 +473,10 @@ impl<
         Factor::has_shortest_path_ambiguity(&Self::factor_params(params))
     }
 
-    fn dparams_matrix(params: &<S>::Vector<PARAMS>, col_idx: u8) -> <S>::Matrix<POINT, PARAMS> {
+    fn dparams_matrix(params: &<S>::Vector<PARAMS>, col_idx: usize) -> <S>::Matrix<POINT, PARAMS> {
         let factor_params = &Self::factor_params(params);
 
-        if col_idx < POINT as u8 {
+        if col_idx < POINT {
             S::Matrix::block_mat1x2::<POINT, SPARAMS>(
                 S::Matrix::zeros(),
                 Factor::dparams_matrix(factor_params, col_idx),
